move Token type to shared-pure
