--- packages/backend/src/model/Token.ts
@@ -1,9 +1,0 @@
-import { AssetId, CoingeckoId, EthereumAddress } from '@l2beat/shared-pure'
-
-export interface Token {
-  id: AssetId
-  symbol: string
-  decimals: number
-  address?: EthereumAddress
-  coingeckoId: CoingeckoId
-}
