+++ packages/backend/src/model/index.ts
@@ -1,2 +1,1 @@
 export * from './Project'
-export * from './Token'
