+++ packages/shared-pure/src/types/index.ts
@@ -9,5 +9,6 @@ export * from './EthereumAddress'
 export * from './Hash256'
 export * from './json'
 export * from './ProjectId'
+export * from './Token'
 export * from './UnixTime'
 export * from './ValueType'
